moved class outside of checkmate
